@@ -1,26 +1,19 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.11'
+gem 'rails', '4.0.0.rc1'
+
+gem 'sass-rails', '~> 4.0.0.rc1'
+gem 'uglifier', '>= 1.3.0'
+gem 'coffee-rails', '~> 4.0.0'
+gem 'therubyracer', platforms: :ruby
+gem 'jquery-rails'
 
 gem 'mysql2'
 
-gem 'bootstrap-sass'
 gem 'kaminari'
 gem 'simple_form'
-gem 'devise'
+gem 'devise', '3.0.0.rc'
 gem 'responders'
-
-# Gems used only for assets and not required
-# in production environments by default.
-group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
-  gem 'therubyracer'
-  gem 'uglifier', '>= 1.0.3'
-end
-
-gem 'jquery-rails'
-gem 'chosen-rails'
 
 group :development do
   gem 'pry-rails'
