@@ -1,7 +1,12 @@
 # Be sure to restart your server when you modify this file.
 
-# Your secret key for verifying the integrity of signed cookies.
+# Your secret key is used for verifying the integrity of signed cookies.
 # If you change this key, all old signed cookies will become invalid!
+
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-BolaDePelo::Application.config.secret_token = 'f9a7443598052e58e25c3e8cd3c69808ac9d6dc6e14a920bd5be4f0ea3001a3ff38a8de64f82f3b073f1ba86ba7194d0237d85f7ba4a8b60a472fdb05834e163'
+# You can use `rake secret` to generate a secure secret key.
+
+# Make sure your secret_key_base is kept private
+# if you're sharing your code publicly.
+BolaDePelo::Application.config.secret_key_base = '633316d50ca8e033d5cce20ecf771d1a35a71b6fe8850883750c3fe017625e6f2fac1fb61ba86bc250a492494cf3ffd21057e5ba5f32e267f79b0e6a9885459d'
